requirements.txt
# Core web framework
streamlit>=1.28.0
flask>=2.3.0
flask-sqlalchemy>=3.0.0
flask-login>=0.6.0
flask-wtf>=1.1.0
flask-cors>=4.0.0

# Database
alembic>=1.12.0

# Data analysis - use latest compatible versions
pandas>=2.0.0
numpy>=1.24.0
numpy-financial>=1.0.0
scipy>=1.11.0
matplotlib>=3.7.0

# PDF and visualization
reportlab>=4.0.0
plotly>=5.15.0

# Security and utilities
passlib>=1.7.0
python-dotenv>=1.0.0
requests>=2.31.0

# Production server
gunicorn>=21.2.0

# Development
pytest>=7.4.0
